requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
docopt==0.6.2
fonttools==4.61.1
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
lsq-ellipse==2.2.1
matplotlib==3.10.8
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
threadpoolctl==3.6.0
urllib3==2.6.3
yarg==0.1.10
